refactor(ingestion): split heritage/bindings/dead_code monoliths into per-language subpackages
- heritage.py (623 LOC) -> extractors/heritage/{lang}.py with __init__.py
  re-exporting extract_heritage, HERITAGE_EXTRACTORS, heritage_node_types_for
- bindings.py (527 LOC) -> extractors/bindings/{lang}.py with __init__.py
  dispatcher preserving extract_import_bindings public API
- dead_code.py (744 LOC) -> analysis/dead_code/ subpackage with analyzer,
  models, constants, and dynamic_markers separated; public API
  (DeadCodeAnalyzer, DeadCodeKind, DeadCodeFindingData, DeadCodeReport)
  preserved

Also fixes Swift extension protocol conformance: extension_declaration was
parsed correctly by _extract_swift_heritage but the upstream filter in
languages/registry.py discarded it. Adds extension_declaration to Swift's
heritage_node_types so 'extension Foo: Protocol {}' produces a heritage
edge.

Adds first regression coverage for Kotlin and Swift extractors
(test_kotlin_extractors.py, test_swift_extractors.py) covering interface
parsing as class_declaration, class/object capture, extension
conformance, multi-protocol conformance, and class/protocol inheritance.
