+++ packages/core/src/repowise/core/analysis/dead_code/models.py
@@ -1,0 +1,42 @@
+"""Public dataclasses + enum for dead-code findings."""
+
+from __future__ import annotations
+
+from dataclasses import dataclass
+from datetime import datetime
+from enum import StrEnum
+
+
+class DeadCodeKind(StrEnum):
+    UNREACHABLE_FILE = "unreachable_file"
+    UNUSED_EXPORT = "unused_export"
+    UNUSED_INTERNAL = "unused_internal"
+    ZOMBIE_PACKAGE = "zombie_package"
+
+
+@dataclass
+class DeadCodeFindingData:
+    kind: DeadCodeKind
+    file_path: str
+    symbol_name: str | None
+    symbol_kind: str | None
+    confidence: float
+    reason: str
+    last_commit_at: datetime | None
+    commit_count_90d: int
+    lines: int
+    package: str | None
+    evidence: list[str]
+    safe_to_delete: bool
+    primary_owner: str | None
+    age_days: int | None
+
+
+@dataclass
+class DeadCodeReport:
+    repo_id: str
+    analyzed_at: datetime
+    total_findings: int
+    findings: list[DeadCodeFindingData]
+    deletable_lines: int
+    confidence_summary: dict  # {"high": N, "medium": N, "low": N}
