+++ packages/core/src/repowise/core/ingestion/extractors/bindings/ts_js.py
@@ -1,0 +1,64 @@
+"""TypeScript / JavaScript import-binding extraction."""
+
+from __future__ import annotations
+
+from tree_sitter import Node
+
+from ...models import NamedBinding
+from ..helpers import node_text
+
+
+def extract_ts_js_bindings(stmt_node: Node, src: str) -> tuple[list[str], list[NamedBinding]]:
+    """Extract bindings from TypeScript/JavaScript import statements."""
+    names: list[str] = []
+    bindings: list[NamedBinding] = []
+
+    for child in stmt_node.children:
+        if child.type != "import_clause":
+            continue
+        for sub in child.children:
+            if sub.type == "identifier":
+                local = node_text(sub, src)
+                names.append(local)
+                bindings.append(
+                    NamedBinding(local_name=local, exported_name="default", source_file=None)
+                )
+            elif sub.type == "named_imports":
+                for spec in sub.children:
+                    if spec.type != "import_specifier":
+                        continue
+                    name_node = spec.child_by_field_name("name") or (
+                        spec.children[0] if spec.children else None
+                    )
+                    alias_node = spec.child_by_field_name("alias")
+                    if name_node:
+                        exported = node_text(name_node, src)
+                        local = node_text(alias_node, src) if alias_node else exported
+                        names.append(local)
+                        bindings.append(
+                            NamedBinding(
+                                local_name=local, exported_name=exported, source_file=None
+                            )
+                        )
+            elif sub.type == "namespace_import":
+                ns_name = None
+                for ns_child in sub.children:
+                    if ns_child.type == "identifier":
+                        ns_name = node_text(ns_child, src)
+                if ns_name:
+                    names.append(ns_name)
+                    bindings.append(
+                        NamedBinding(
+                            local_name=ns_name,
+                            exported_name=None,
+                            source_file=None,
+                            is_module_alias=True,
+                        )
+                    )
+                else:
+                    names.append("*")
+                    bindings.append(
+                        NamedBinding(local_name="*", exported_name=None, source_file=None)
+                    )
+
+    return names, bindings
